simple_agent/pyproject.toml
[project]
name = "simple-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.1.0",
    "langchain-google-genai>=3.1.0",
    "langgraph>=1.0.3",
    "python-dotenv>=1.2.1",
]
